requirements.txt
scikit-learn
streamlit
pandas
pillow